pyproject.toml
[project]
name = "cvex"
version = "0.1.0"
description = "CVEX is a framework for the reproducible exploitation of CVE vulnerabilities"
authors = [
    {name = "Oleg Boyarchuk", email = "oleg.boyarchuk@broadcom.com"},
    {name = "Gabriel Pizarro", email = "gpizarro@ucsb.edu"}
]
readme = "README.md"

requires-python = ">=3.10"

dependencies = [
  "python-vagrant >= 1.0.0",
  "paramiko >= 3.4.1",
  "PyYAML >= 6.0.2",
  "fabric >= 3.2.2",
  "procmon-parser >= 0.3.13"
]

[project.scripts]
cvex = "cvex.cvex:main"

